+++ Sources/DoggieGPU/CoreImage/Shaders/SVGDisplacementMapKernel.metal
@@ -31,16 +31,16 @@ constexpr sampler linear_sampler (coord::pixel, address::clamp_to_edge, filter::
 constant int X_SELECTOR [[function_constant(0)]];
 constant int Y_SELECTOR [[function_constant(1)]];
 
-kernel void svg_displacement_map(texture2d<half, access::sample> source [[texture(0)]],
-                                 texture2d<half, access::read> displacement [[texture(1)]],
-                                 texture2d<half, access::write> output [[texture(2)]],
+kernel void svg_displacement_map(texture2d<float, access::sample> source [[texture(0)]],
+                                 texture2d<float, access::read> displacement [[texture(1)]],
+                                 texture2d<float, access::write> output [[texture(2)]],
                                  constant packed_float2 &offset [[buffer(3)]],
                                  constant packed_float2 &scale [[buffer(4)]],
                                  uint2 gid [[thread_position_in_grid]]) {
     
     if (gid.x >= output.get_width() || gid.y >= output.get_height()) { return; }
     
-    const half4 d = displacement.read(gid);
+    const float4 d = displacement.read(gid);
     
     const float _x = d[X_SELECTOR] - 0.5;
     const float _y = d[Y_SELECTOR] - 0.5;
@@ -49,7 +49,7 @@ kernel void svg_displacement_map(texture2d<half, access::sample> source [[textur
     coord.x += _x * scale[0];
     coord.y -= _y * scale[1];
     
-    const half4 color = source.sample(linear_sampler, coord);
+    const float4 color = source.sample(linear_sampler, coord);
     
     output.write(color, gid);
 }
